Mostly Complete (ver. 1)
Reminders pop up on time, and the UI has been improved.  It's possible
to edit reminders and delete them.

Needs work: The notifications should have a sound & accompanied action
when touched.  The application needs to be able to handle when the OS
takes away its resources when it's in the background